Add files via upload
This is code to extract Global Field Power (GFP), and run a repeated-measures cluster permutation ANOVA to test for differences of GFP between conditions (the ANOVA within the permutation cluster test can be changed, setting stats to "None" will result in a simple one-way ANOVA to be used. The code first loads libraries, then creates variables and loads the data (for each subject and condition), then extracts the GFP for each subject and condition (of the average ERP for that subject and condition), and then runs a cluster-permutation test. Further instructions are contained in the code.